Revert "mkfs.f2fs: support <1% overprovision ratio"
This reverts commit ee5c82f49d49308bb17423ab170792cd3c302c8e.
It fix crash mkfs.f2fs on lot of msm devices.

+++ mkfs/f2fs_format_main.c
@@ -101,7 +101,7 @@ static void f2fs_parse_options(int argc, char *argv[])
 			config.vol_label = optarg;
 			break;
 		case 'o':
-			config.overprovision = atof(optarg);
+			config.overprovision = atoi(optarg);
 			break;
 		case 'O':
 			parse_feature(strdup(optarg));
